requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Django==5.0.6
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
